+++ docs/404.html
@@ -1,0 +1,7 @@
+<!DOCTYPE html><head><meta charset="utf-8"><meta name="viewport" content="width=device-width,initial-scale=1"><meta name="generator" content="Compositor x0"><title>React Website Starter</title><meta name="description" content="Simple setup for building sites with React"><link rel="icon" href="assets/favicon.png"><link rel="apple-touch-icon" href="assets/apple-touch-icon.png"><style>*{box-sizing:border-box}body{margin:0;font-family:system-ui,sans-serif}</style><style data-styled-components="dUiVDW iuMQGz">/* sc-component-id: sc-bxivhb */
+.dUiVDW{font-family:system-ui,sans-serif;} .dUiVDW *{box-sizing:border-box;}
+/* sc-component-id: sc-iELTvK */
+.iuMQGz{color:#24292e;}</style></head><div id="root"><div class="sc-bxivhb dUiVDW"><style>body {
+      font-family: -apple-system, BlinkMacSystemFont, "Segoe UI", Helvetica, Arial, sans-serif;
+      line-height: 1.5;
+    }</style><div class="sc-iELTvK iuMQGz" color="bodytext"><pre>/Users/jxnblk/repos/react-website-starter/pages</pre><ul><li><a href="/">index</a></li></ul></div></div></div>
