+++ docs/index.html
@@ -1,0 +1,9 @@
+<!DOCTYPE html><head><meta charset="utf-8"><meta name="viewport" content="width=device-width,initial-scale=1"><meta name="generator" content="Compositor x0"><title>React Website Starter</title><meta name="description" content="Simple setup for building sites with React"><link rel="icon" href="assets/favicon.png"><link rel="apple-touch-icon" href="assets/apple-touch-icon.png"><style>*{box-sizing:border-box}body{margin:0;font-family:system-ui,sans-serif}</style><style data-styled-components="dUiVDW iuMQGz cEvLkB jTacKi">/* sc-component-id: sc-bxivhb */
+.dUiVDW{font-family:system-ui,sans-serif;} .dUiVDW *{box-sizing:border-box;}
+/* sc-component-id: sc-iELTvK */
+.iuMQGz{color:#24292e;}.cEvLkB{margin-left:auto;margin-right:auto;padding:16px;width:100%;} @media screen and (min-width:544px){.cEvLkB{width:100%;}} @media screen and (min-width:768px){.cEvLkB{width:66.66666666666666%;}}
+/* sc-component-id: sc-fYxtnH */
+.jTacKi{margin:0px;font-size:32px;font-weight:600;line-height:1.25;}</style></head><div id="root"><div class="sc-bxivhb dUiVDW"><style>body {
+      font-family: -apple-system, BlinkMacSystemFont, "Segoe UI", Helvetica, Arial, sans-serif;
+      line-height: 1.5;
+    }</style><div class="sc-iELTvK iuMQGz" color="bodytext"><div class="sc-iELTvK cEvLkB" width="1,1,0.6666666666666666"><h1 class="sc-fYxtnH jTacKi" font-size="5" font-weight="600">Hello world</h1></div></div></div></div>
